wasi: Move fizzy-wasi main() to separate file
