+++ tools/wasi/CMakeLists.txt
@@ -1,14 +1,12 @@
 # Fizzy: A fast WebAssembly interpreter
-# Copyright 2019-2020 The Fizzy Authors.
+# Copyright 2020 The Fizzy Authors.
 # SPDX-License-Identifier: Apache-2.0
 
 include(ProjectUVWASI)
 
-set(fizzy_include_dir ${PROJECT_SOURCE_DIR}/lib/fizzy)
-
-add_executable(fizzy-wasi wasi.cpp)
+add_executable(fizzy-wasi)
 target_link_libraries(fizzy-wasi PRIVATE fizzy::fizzy-internal uvwasi::uvwasi)
-target_include_directories(fizzy-wasi PRIVATE ${fizzy_include_dir})
+target_sources(fizzy-wasi PRIVATE main.cpp wasi.cpp wasi.hpp)
 
 if(UNIX AND NOT APPLE)
     # For libstdc++ up to version 8 (included) this is needed for proper <filesystem> support.
